Fixed issues with scripts, finished animations, added comments to scripts.
Fixed an issue with player sliding down slopes when idle. This was fixed by freezing the x position on the rigidbody.
Fixed an issue with camerafollow stuttering when following. This was fixed by multiplying the new camera transform by Time.deltatime
Fixed an issue where the player will not jump when the jump key is held down. This was fixed by using the same logic for instant un-crouch by using Input.GetButtonUp(jumpbutton)

Created a bounding box that will not let the player fall out of the map.
Updated README.

Finished the animations.
There are still a couple glitches where the animation will cut before it's completely done but for not it's working decently.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>