Started on stats, created collectables, updated the environment
The collectables are items but for now there is only a pickup animation but no actual logic to it other than a respawn timer.
The environment was lacking some background objects so I've added them in now.
I've moved the animated placeholder objects further away to remove unneeded distraction until they're ready to be implemented.
Updated the readme.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>