Updated the UI, changed scripts
The UI now uses Pixel Art.
The Slider will now display 0 fill when the player's health is below 0.

In progress:
Update EntityStats to have all variables with full getters/setters.
Update UIManager to have all the variables required to display the stats on the UI.
Update the UI to contain the Stat window.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>